refactor: move format function to lib folder
